CotC: Syslog and LTTng tracing support
The Calvin log maybe routed to local or networked syslog using --logfile or -f
and the specification syslog[:Location]. Location may be a device path or a
network address. The latter may contain an additional colon, ':', followed by
a port number.

Trace support is implemented using Cython. The preferred method is to use LTTng
tracing and collect node traces using an LTTng relay. There is lingering code
which uses an internal trace buffer and stores CTF when the Calvin runtime is
terminated but it has not been kept up to date.